backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
django-graphql-jwt==0.3.4
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
Pillow==10.1.0
promise==2.3
PyJWT==2.8.0
pytz==2023.3.post1
Rx==1.6.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.8.0
